fix(editor): Move project store
