+++ packages/editor-ui/src/init.ts
@@ -6,7 +6,7 @@ import { useSourceControlStore } from '@/stores/sourceControl.store';
 import { useUsersStore } from '@/stores/users.store';
 import { initializeCloudHooks } from '@/hooks/register';
 import { useVersionsStore } from '@/stores/versions.store';
-import { useProjectsStore } from '@/features/projects/projects.store';
+import { useProjectsStore } from '@/stores/projects.store';
 import { useRolesStore } from './stores/roles.store';
 
 let coreInitialized = false;
